A few cosmetic repairs.

+++ run_io_avstats.sh
@@ -150,6 +150,9 @@ if [ "${IO_AERO_TASK}" = "l_c_d" ]; then
     fi
 fi
 
+# Path to the log file
+log_file="run_io_avstats_${IO_AERO_TASK}.log"
+
 # Function for logging messages
 log_message() {
   local message="$1"
@@ -159,17 +162,27 @@ log_message() {
   echo "$timestamp: $message" >> "$log_file"
 }
 
+# Check if logging_io_aero.log exists and delete it
+if [ -f logging_io_aero.log ]; then
+    rm -f logging_io_aero.log
+fi
+
+# Check if the file specified in logfile exists and delete it
+if [ -f "${log_file}" ]; then
+    rm -f "${log_file}"
+fi
+
 # Redirection of the standard output and the standard error output to the log file
 exec > >(while read -r line; do log_message "$line"; done) 2> >(while read -r line; do log_message "ERROR: $line"; done)
-
-rm -f logging_io_aero.log
 
 echo "================================================================================"
 echo "Start $0"
 echo "--------------------------------------------------------------------------------"
 echo "IO-AVSTATS - Aviation Event Statistics."
 echo "--------------------------------------------------------------------------------"
-echo "PYTHONPATH   : ${PYTHONPATH}"
+echo "ENV_FOR_DYNACONF         : ${ENV_FOR_DYNACONF}"
+echo "POSTGRES_CONNECTION_PORT : ${IO_AERO_POSTGRES_CONNECTION_PORT}"
+echo "PYTHONPATH               : ${PYTHONPATH}"
 echo "--------------------------------------------------------------------------------"
 echo "TASK         : ${IO_AERO_TASK}"
 echo "APPLICATION  : ${IO_AERO_APPLICATION}"
